Rename android package
